refactor: move iMessage channel to extensions/imessage (#45539)
(cherry picked from commit 0ce23dc62d376f5625a3c6c572d07d8cac0c16dc)
