+++ extensions/imessage/src/monitor.ts
@@ -1,0 +1,2 @@
+export { monitorIMessageProvider } from "./monitor/monitor-provider.js";
+export type { MonitorIMessageOpts } from "./monitor/types.js";
