+++ src/imessage/constants.ts
@@ -1,2 +1,2 @@
-/** Default timeout for iMessage probe/RPC operations (10 seconds). */
-export const DEFAULT_IMESSAGE_PROBE_TIMEOUT_MS = 10_000;
+// Shim: re-exports from extensions/imessage/src/constants
+export * from "../../extensions/imessage/src/constants.js";
