+++ src/imessage/monitor.ts
@@ -1,2 +1,2 @@
-export { monitorIMessageProvider } from "./monitor/monitor-provider.js";
-export type { MonitorIMessageOpts } from "./monitor/types.js";
+// Shim: re-exports from extensions/imessage/src/monitor
+export * from "../../extensions/imessage/src/monitor.js";
